fix HomeView UI
Remove the openGames part of the UI, because I am not implementing that
any time soon.
Change the UI so that it only asks for player 2's public key to create a game,
or join an existing game by entering the game's public key.

TODO: Fix walletStore not being set by initWallet() from solana-wallets-vue