remove wallet init functions from main.ts
This commit moves all the wallet init functions to workspace.ts, because otherwise,
none of the other components will be able to see the wallet that's connected.

For some reason, using SolanaWallets as a plugin doesn't seem to work either,
and the wallet isn't visible to any of the other components.