add createWorkspace to workspace.ts
This commit adds a function that returns an object containing
the Solana program's idl, solana web3 connection, current connected wallet, and
anchor provider.
This is because other components need to import the workspace themselves. Using
useWallet() or useAnchorWallet() on their own doesn't work. I can't be bothered to
figure it out anymore, so this will be a hack.